Expand the index encoding/decoding routines to support any unsigned type of at least 2 bytes.

+++ engine/utilities/tightencoding.cpp
@@ -67,10 +67,13 @@ template Integer tightDecodingInteger<Integer, it2>(it2, it2, bool);
 template LargeInteger tightDecodingInteger<LargeInteger, it1>(it1, it1, bool);
 template LargeInteger tightDecodingInteger<LargeInteger, it2>(it2, it2, bool);
 
-template void tightEncodeIndex<size_t>(std::ostream&, size_t);
+// End users don't get to use the index routines, so what we instantiate
+// here simply needs to be whatever Regina uses internally.
+
+template void tightEncodeIndex<unsigned long>(std::ostream&, unsigned long);
 template void tightEncodeIndex<ssize_t>(std::ostream&, ssize_t);
 
-template size_t tightDecodingIndex<size_t>(std::istream&);
+template unsigned long tightDecodingIndex<unsigned long>(std::istream&);
 template ssize_t tightDecodingIndex<ssize_t>(std::istream&);
 
 } // namespace regina::detail
